+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.1.0](https://github.com/jorggerojas/create-skeleton-next/compare/v1.0.1...v1.1.0) (2026-02-24)
+
+
+### Features
+
+* **core,tests:** add package manager options ([4b10b4b](https://github.com/jorggerojas/create-skeleton-next/commit/4b10b4b845e6efb7a9418f324c412f71b062a24f))
+* **notifier:** notify package updates ([66d27c4](https://github.com/jorggerojas/create-skeleton-next/commit/66d27c4bdbed263af532ab4b4a1af37d9595db54))
+
+
+### Bug Fixes
+
+* **all:** define import paths instead of relative ones ([009a115](https://github.com/jorggerojas/create-skeleton-next/commit/009a1150c4e04057e79b123326bc80f02c4663e9))
+
+
+### Documentation
+
+* update readme file ([cb6ce3e](https://github.com/jorggerojas/create-skeleton-next/commit/cb6ce3e52b6291d06c9e607b0c6d2acdab6d2ac3))
+
 ### [1.0.1](https://github.com/jorggerojas/create-skeleton-next/compare/v1.0.0...v1.0.1) (2026-02-02)
 
 ## [1.0.0](https://github.com/jorggerojas/create-skeleton-next/compare/v0.1.0...v1.0.0) (2026-02-02)